# Use by default generic lu factorization since it is faster.