--- a/src/correctors/newton.jl
+++ b/src/correctors/newton.jl
@@ -22,7 +22,7 @@ end
 function correct!(xnext, ::Newton, cache::NewtonCache, H::HomotopyWithCache, x, t, tol, maxiters)
     A, b = cache.A, cache.b
     evaluate_and_jacobian!(b, A, H, x, t)
-    solve_with_lu_inplace!(A, b)
+    ldiv_lu!(A, b)
     @. xnext = x - b
     k = 1
     while true
@@ -39,7 +39,7 @@ function correct!(xnext, ::Newton, cache::NewtonCache, H::HomotopyWithCache, x, 
         # put jacobian in A
         jacobian!(A, H, xnext, t)
 
-        solve_with_lu_inplace!(A, b)
+        ldiv_lu!(A, b)
         @. xnext = xnext - b
     end
 end