@saschatimme 2018-05-16
I wanted to add the old version as well, but always hit a compiler bug similar to [#22355](https://github.com/JuliaLang/julia/issues/22355), so this is out for now.